requirements.txt
cv2
numpy
imutils
piServoCtl
RPIO